refactor: switch python binds unit tests to `pytest`
