+++ README.md
@@ -54,6 +54,7 @@ cmake --build . --parallel 4
 Our Python binds target Python 3.10 or higher; they depend on
 
 * [`pybind11`](https://github.com/pybind/pybind11) (bridging C++ and Python)
+* [`pytest`](https://github.com/pytest-dev/pytest) (unit tests, part of optional `[.dev]` dependency group)
 * [`ruff`](https://github.com/astral-sh/ruff) (linting, part of optional `[.dev]` dependency group)
 
 > [!NOTE]
@@ -78,7 +79,7 @@ uv run ruff check
 uv run ruff format --check
 
 # Run unit tests
-uv run python binds/python/test.py
+uv run pytest -v binds/python/test_unit.py
 
 # Run benchmarks
 uv run python binds/python/benchmark.py
