+++ pyproject.toml
@@ -17,6 +17,7 @@ Issues = "https://github.com/dashpay/bls-signatures/issues"
 [project.optional-dependencies]
 dev = [
   "ruff>=0.9",
+  "pytest>=8.1",
 ]
 
 [build-system]
@@ -47,7 +48,7 @@ select = [
 ]
 
 [tool.ruff.lint.per-file-ignores]
-"binds/python/test.py" = ["S101"]
+"binds/python/test*.py" = ["S101"]
 
 [tool.ruff.format]
 indent-style = "space"
